[特性新增] #time 8h 补齐log_reader,log_writer缺失的crc校验

+++ src/util/coding.h
@@ -23,8 +23,8 @@ uint8_t const *GetVarint64Ptr(uint8_t const *p, uint8_t const *limit, uint64_t *
 
 int32_t VarintLength(uint64_t value);
 
-inline uint64_t DecodeFixed32(uint8_t const *ptr) {
-    return (static_cast<uint64_t>(ptr[0])) | (static_cast<uint64_t>(ptr[1]) << 8) | (static_cast<uint64_t>(ptr[2]) << 16) | (static_cast<uint64_t>(ptr[3]) << 24);
+inline uint32_t DecodeFixed32(uint8_t const *ptr) {
+    return (static_cast<uint32_t>(ptr[0])) | (static_cast<uint32_t>(ptr[1]) << 8) | (static_cast<uint32_t>(ptr[2]) << 16) | (static_cast<uint32_t>(ptr[3]) << 24);
 }
 inline uint64_t DecodeFixed64(uint8_t const *ptr) {
     return (static_cast<uint64_t>(ptr[0])) | (static_cast<uint64_t>(ptr[1]) << 8) | (static_cast<uint64_t>(ptr[2]) << 16) | (static_cast<uint64_t>(ptr[3]) << 24) | (static_cast<uint64_t>(ptr[4]) << 32) | (static_cast<uint64_t>(ptr[5]) << 40) | (static_cast<uint64_t>(ptr[6]) << 48) | (static_cast<uint64_t>(ptr[7]) << 56);
